% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcem.R
\name{covfndef}
\alias{covfndef}
\title{define covariance function}
\usage{
covfndef(nu)
}
\arguments{
\item{nu}{Matern covariance function smoothness. select from 0.5,1.5,2.5,10}
}
\value{
covariance function
}
\description{
define covariance function
}
