% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcem.R
\name{sparse.sglmmGP.mcem.pred}
\alias{sparse.sglmmGP.mcem.pred}
\title{Fit MCEM to projection-based SGLMM}
\usage{
sparse.sglmmGP.mcem.pred(fit, pred.X, pred.coords, m = 1000,
  burn = 0.5, rdist = rdist)
}
\arguments{
\item{Z}{a vector of observation data}

\item{x}{a matrix storing the covariates including the intercept}

\item{coords}{data location}

\item{family}{"poisson"}
}
\value{
beta.est regresson coefficient estimates

wmean.update random effects estimates
}
\description{
This function runs the Markov chain Monte Carlo EM algorithm
to fit SGLMM using the projection-based approach. The method paper
Fast expectation-maximization algorithms for spatial generalized linear mixed models
can be found here https://arxiv.org/abs/1909.05440
}
