% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcem.R
\name{sparse.sglmmGP.mcem}
\alias{sparse.sglmmGP.mcem}
\title{Fit MCEM to projection-based SGLMM}
\usage{
sparse.sglmmGP.mcem(Z, X, coords, family = "poisson", offset = NULL,
  q = 50, MCN = 30, nu, covfn = NULL, mc.cores = 1, mul = 2,
  init = NULL, ceil = 500, epsilon = 0.01, zalpha = qnorm(0.75),
  ifbreak = T, RSR = T, RP = F, tune = 0.1, track = T, ALL = T)
}
\arguments{
\item{Z}{a vector of observation data}

\item{coords}{data location}

\item{family}{"poisson"}

\item{nu}{Matern covariance function smoothness. select from 0.5,1.5,2.5,10}

\item{x}{a matrix storing the covariates including the intercept}
}
\value{
beta.est regresson coefficient estimates

wmean.update random effects estimates
}
\description{
This function runs the Markov chain Monte Carlo EM algorithm
to fit SGLMM using the projection-based approach. The method paper
Fast expectation-maximization algorithms for spatial generalized linear mixed models
can be found here https://arxiv.org/abs/1909.05440
}
