% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcem.R
\name{sparse.sglmmGP.mcem.pred}
\alias{sparse.sglmmGP.mcem.pred}
\title{Fit MCEM to projection-based SGLMM}
\usage{
sparse.sglmmGP.mcem.pred(fit, pred.X, pred.coords, m = 1000,
  burn = 0.5)
}
\arguments{
\item{fit}{output from sparse.sglmmGP.mcem}

\item{pred.X}{a matrix storing the covariates at prediction loc}

\item{pred.coords}{a matrix prediction loc}

\item{family}{}
}
\value{
pred.mean a matrix storing 1000 draw of prediction
}
\description{
This function runs the Markov chain Monte Carlo EM algorithm
to fit SGLMM using the projection-based approach. The method paper
Fast expectation-maximization algorithms for spatial generalized linear mixed models
can be found here https://arxiv.org/abs/1909.05440
}
